Replace signed long with (unsigned) size_t

+++ src/qt/clientmodel.h
@@ -53,7 +53,7 @@ public:
     int getNumBlocks() const;
 
     //! Return number of transactions in the mempool
-    long getMempoolSize() const;
+    size_t getMempoolSize() const;
     //! Return the dynamic memory usage of the mempool
     size_t getMempoolDynamicUsage() const;
     
@@ -90,7 +90,7 @@ private:
 Q_SIGNALS:
     void numConnectionsChanged(int count);
     void numBlocksChanged(int count, const QDateTime& blockDate, double nVerificationProgress);
-    void mempoolSizeChanged(long count, size_t mempoolSizeInBytes);
+    void mempoolSizeChanged(size_t count, size_t mempoolSizeInBytes);
     void alertsChanged(const QString &warnings);
     void bytesChanged(quint64 totalBytesIn, quint64 totalBytesOut);
